pyproject.toml
[project]
name = "grocyscan"
version = "0.3.4-alpha"
description = "Barcode scanning app for Grocy inventory management"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "GrocyScan Team" }]
dependencies = [
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "nicegui>=1.4.0",
    
    # Database
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",
    
    # Cache
    "redis>=5.0.0",
    
    # Validation & Settings
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # HTTP Client
    "httpx>=0.26.0",
    
    # LLM
    "litellm>=1.20.0",
    
    # Authentication
    "bcrypt>=4.1.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    
    # Logging & Observability
    "structlog>=24.1.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-exporter-otlp>=1.22.0",
    "opentelemetry-instrumentation-fastapi>=0.43b0",
    "opentelemetry-instrumentation-httpx>=0.43b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.43b0",
    "prometheus-client>=0.19.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "rapidfuzz>=3.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "ruff>=0.1.0",
    "black>=24.1.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "I", "N", "UP", "ANN", "B", "C4", "SIM"]
ignore = ["ANN101", "ANN102", "ANN401"]

[tool.ruff.isort]
known-first-party = ["app"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
